+++ lib/wrapture/class_spec.rb
@@ -40,6 +40,7 @@ module Wrapture
 
       normalized['version'] = Wrapture.spec_version(spec)
       normalized['includes'] = Wrapture.normalize_includes(spec['includes'])
+      normalized['type'] = ClassSpec.effective_type(normalized)
 
       if spec.key?('parent')
         includes = Wrapture.normalize_includes(spec['parent']['includes'])
@@ -47,6 +48,27 @@ module Wrapture
       end
 
       normalized
+    end
+
+    # Gives the effective type of the given class spec hash.
+    def self.effective_type(spec)
+      inferred_pointer_wrapper = spec['constructors'].any? do |func|
+        func['wrapped-function']['return']['type'] == EQUIVALENT_POINTER_KEYWORD
+      end
+
+      if spec.key?('type')
+        valid_types = %w[pointer struct]
+        unless valid_types.include?(spec['type'])
+          type_message = "#{spec['type']} is not a valid class type"
+          raise InvalidSpecKey.new(type_message, valid_keys: valid_types)
+        end
+
+        spec['type']
+      elsif inferred_pointer_wrapper
+        'pointer'
+      else
+        'struct'
+      end
     end
 
     # Returns a string of the variable with it's type, properly formatted.
@@ -466,9 +488,7 @@ module Wrapture
 
     # Determines if this class is a wrapper for a struct pointer or not.
     def pointer_wrapper?
-      @spec['constructors'].any? do |spec|
-        spec['wrapped-function']['return']['type'] == EQUIVALENT_POINTER_KEYWORD
-      end
+      @spec['type'] == 'pointer'
     end
 
     # The signature of the constructor given an equivalent struct type.